Remove JsonSchema generation
