+++ .github/workflows/update-config.yaml
@@ -1,4 +1,4 @@
-name: "Update Config Schema"
+name: "Update Config Types"
 
 permissions:
   contents: write
@@ -20,7 +20,7 @@ jobs:
       - uses: actions/checkout@v4
       - uses: actions-rust-lang/setup-rust-toolchain@v1
 
-      - run: cargo run --bin regenerate-config-schema
+      - run: cargo run --bin regenerate-config-types
 
       - name: Commit and push changes
         run: |
@@ -30,7 +30,7 @@ jobs:
           if git diff --cached --quiet; then
             echo "No changes to commit"
           else
-            git commit -m 'Update config_schema.json'
+            git commit -m 'Update config Lua types'
             git push
           fi
         env:
