+++ config/Cargo.toml
@@ -7,9 +7,7 @@ edition = "2021"
 anyhow = "1.0.79"
 clap = { version = "4.5.3", features = ["derive"] }
 latest_bin = { path = "../latest_bin" }
-schemars = "0.8.21"
 serde = { version = "1.0.197", features = ["derive"] }
-serde_json = "1.0.122"
 tracing = "0.1.40"
 tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
 mlua = { version = "0.9.7", features = ["lua54", "vendored", "serialize"] }
