--- config/config_schema.json
@@ -1,148 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "Config",
-  "description": "Configuration for the application.",
-  "type": "object",
-  "properties": {
-    "shell_caching": {
-      "description": "Optional configuration for cache-shell-setup",
-      "anyOf": [
-        {
-          "$ref": "#/definitions/ShellCache"
-        },
-        {
-          "type": "null"
-        }
-      ]
-    },
-    "tmux": {
-      "description": "Optional tmux configuration. Including sessions and windows to be created.",
-      "anyOf": [
-        {
-          "$ref": "#/definitions/Tmux"
-        },
-        {
-          "type": "null"
-        }
-      ]
-    }
-  },
-  "definitions": {
-    "Command": {
-      "description": "Command to be executed in a tmux window.",
-      "anyOf": [
-        {
-          "description": "A single command as a string.",
-          "type": "string"
-        },
-        {
-          "description": "Multiple commands as a list of strings.",
-          "type": "array",
-          "items": {
-            "type": "string"
-          }
-        }
-      ]
-    },
-    "Session": {
-      "description": "Configuration for a tmux session.",
-      "type": "object",
-      "required": [
-        "name",
-        "windows"
-      ],
-      "properties": {
-        "name": {
-          "description": "Name of the session.",
-          "type": "string"
-        },
-        "windows": {
-          "description": "List of windows in the session.",
-          "type": "array",
-          "items": {
-            "$ref": "#/definitions/Window"
-          }
-        }
-      }
-    },
-    "ShellCache": {
-      "type": "object",
-      "required": [
-        "destination",
-        "source"
-      ],
-      "properties": {
-        "destination": {
-          "type": "string"
-        },
-        "source": {
-          "type": "string"
-        }
-      }
-    },
-    "Tmux": {
-      "description": "Tmux configuration.",
-      "type": "object",
-      "required": [
-        "sessions"
-      ],
-      "properties": {
-        "default_session": {
-          "description": "The default session to attach to when `startup-tmux --attach` is ran.",
-          "type": [
-            "string",
-            "null"
-          ]
-        },
-        "sessions": {
-          "description": "List of tmux sessions.",
-          "type": "array",
-          "items": {
-            "$ref": "#/definitions/Session"
-          }
-        }
-      }
-    },
-    "Window": {
-      "description": "Configuration for a tmux window.",
-      "type": "object",
-      "required": [
-        "name"
-      ],
-      "properties": {
-        "command": {
-          "description": "Optional command to run in the window.",
-          "anyOf": [
-            {
-              "$ref": "#/definitions/Command"
-            },
-            {
-              "type": "null"
-            }
-          ]
-        },
-        "env": {
-          "description": "Additional environment variables to set in the window.",
-          "type": [
-            "object",
-            "null"
-          ],
-          "additionalProperties": {
-            "type": "string"
-          }
-        },
-        "name": {
-          "description": "Name of the window.",
-          "type": "string"
-        },
-        "path": {
-          "description": "Optional path to set as the working directory for the window.",
-          "type": [
-            "string",
-            "null"
-          ]
-        }
-      }
-    }
-  }
-}
